refactor(test): Extract helper to add dummy dim
